Use static_cast for Write_packet payload size

+++ test/rw_packets_test.cpp
@@ -426,7 +426,7 @@ TEST(Write_packet, write_payload_constructor) {
   auto size = sv.size();
   ASSERT_EQ(size, 5u);
 
-  Write_packet p('a', data, size);
+  Write_packet p('a', data, static_cast<std::uint16_t>(size));
   assert_non_empty(p, 'a', size, size, data);
 }
 
@@ -449,7 +449,7 @@ TEST(Write_packet, move_constructor) {
     auto size = sv.size();
     ASSERT_EQ(size, 5u);
 
-    Write_packet p1('a', data, size);
+    Write_packet p1('a', data, static_cast<std::uint16_t>(size));
     Write_packet p2(std::move(p1));
     assert_empty(p1);
     assert_non_empty(p2, 'a', size, size, data);
@@ -477,7 +477,7 @@ TEST(Write_packet, move_assignment) {
     auto size = sv.size();
     ASSERT_EQ(size, 5u);
 
-    Write_packet p1('a', data, size);
+    Write_packet p1('a', data, static_cast<std::uint16_t>(size));
     Write_packet p2;
     p2 = std::move(p1);
     assert_empty(p1);
@@ -490,7 +490,7 @@ TEST(Write_packet, move_assignment) {
     ASSERT_EQ(size, 11u);
 
     Write_packet p1;
-    Write_packet p2('b', data, size);
+    Write_packet p2('b', data, static_cast<std::uint16_t>(size));
     p2 = std::move(p1);
     assert_empty(p1);
     assert_empty(p2);
@@ -503,9 +503,9 @@ TEST(Write_packet, move_assignment) {
 
     std::string_view distinct_initial_value = "HELLO WORLD";
 
-    Write_packet p1('a', data, size);
+    Write_packet p1('a', data, static_cast<std::uint16_t>(size));
     Write_packet p2('b', distinct_initial_value.data(),
-                    distinct_initial_value.size());
+                    static_cast<std::uint16_t>(distinct_initial_value.size()));
     p2 = std::move(p1);
     assert_empty(p1);
     assert_non_empty(p2, 'a', size, size, data);
@@ -531,7 +531,7 @@ TEST(Write_packet, clone) {
     auto size = sv.size();
     ASSERT_EQ(size, 5u);
 
-    Write_packet p1('a', data, size);
+    Write_packet p1('a', data, static_cast<std::uint16_t>(size));
     Write_packet p2 = p1.clone();
     assert_non_empty(p1, 'a', size, size, data);
     assert_non_empty(p2, 'a', size, size, data);
@@ -550,7 +550,7 @@ TEST(Write_packet, clone) {
     auto size = sv.size();
     ASSERT_EQ(size, 11u);
 
-    Write_packet p1('a', data, size);
+    Write_packet p1('a', data, static_cast<std::uint16_t>(size));
     constexpr auto new_size = 5;
     ASSERT_EQ(p1.resize_payload(new_size),
               Write_packet::Resize_result::resized);
@@ -587,13 +587,14 @@ TEST(Write_packet, resize_payload) {
     auto size = sv.size();
     ASSERT_EQ(size, 11u);
 
-    Write_packet p('a', data, size);
+    Write_packet p('a', data, static_cast<std::uint16_t>(size));
     constexpr auto new_size = 5;
     ASSERT_EQ(p.resize_payload(new_size), Write_packet::Resize_result::resized);
     assert_non_empty(p, 'a', size, new_size, data);
     ASSERT_EQ(p.resize_payload(0), Write_packet::Resize_result::resized);
     assert_non_empty(p, 'a', size, 0);
-    ASSERT_EQ(p.resize_payload(size), Write_packet::Resize_result::resized);
+    ASSERT_EQ(p.resize_payload(static_cast<std::uint16_t>(size)),
+              Write_packet::Resize_result::resized);
     assert_non_empty(p, 'a', size, size, data);
   }
 }
@@ -604,7 +605,7 @@ TEST(Write_packet, payload) {
   auto size = sv.size();
   ASSERT_EQ(size, 5u);
 
-  Write_packet p('a', size);
+  Write_packet p('a', static_cast<std::uint16_t>(size));
   std::memcpy(p.payload_data(), data, size);
   assert_non_empty(p, 'a', size, size, data);
 }
